refactor(server): decompose brainstorm send_message and split agent tooling (#607)
Pull brainstorm.py back under the 1k health guard (1064 → 874) by separating
the two concerns the file mixed and decomposing the ~210-line send_message.

- Extract brainstormer agent prompts + restricted filesystem tooling
  (write_design_doc tool, BrainstormerFilesystemMiddleware) into a new
  brainstormer_agent.py module; the service now imports what it needs.
- Decompose send_message at three seams: _resolve_plan_path,
  _detect_artifact_event, _build_complete_event.
- Dedup the driver-cleanup try/except (delete_session +
  update_session_status) into _cleanup_driver_session.
- Dedup the sync/async write_design_doc result handling into
  _design_doc_result.
- Drop 5 toothless tests (constant-exists assertions, removed-feature
  regression checks) covered by sibling content/behavior tests.

Behavior unchanged; full suite (incl. integration) green.

Part of #603
